refactor(xapiri): extract updateDeployTab + renderDeployTab → tab_deploy.go
Mechanical move per ADR 0014. No behaviour change.
